+++ tests/Integration/Common/SpanAssertionTrait.php
@@ -2,6 +2,8 @@
 
 namespace DDTrace\Tests\Integration\Common;
 
+use DDTrace\Span;
+use PHPUnit\Framework\TestCase;
 
 trait SpanAssertionTrait
 {
@@ -15,5 +17,27 @@ trait SpanAssertionTrait
     public function assertExpectedSpans($testCase, $traces, $expectedSpans)
     {
         (new SpanChecker($testCase))->assertSpans($traces, $expectedSpans);
+    }
+
+    /**
+     * Checks that the provide span exists in the provided traces and matches expectations.
+     *
+     * @param TestCase $testCase
+     * @param Span[][] $traces
+     * @param SpanAssertion $expectedSpan
+     */
+    public function assertOneExpectedSpan($testCase, $traces, SpanAssertion $expectedSpan)
+    {
+        $spanChecker = new SpanChecker($testCase);
+
+        $found = array_filter($spanChecker->flattenTraces($traces), function (Span $span) use ($expectedSpan) {
+            return $span->getOperationName() === $expectedSpan->getOperationName();
+        });
+
+        if (empty($found)) {
+            $testCase->fail('Span not found in traces: ' . $expectedSpan->getOperationName());
+        } else {
+            $spanChecker->assertSpan($found[0], $expectedSpan);
+        }
     }
 }